Working STOP mode, low power OK but not great
Whole board down to 4.4mA when ESP8266 is asleep. Odd thing is this drops
to 2.1mA immediatel after STM32 is reset.

Cause appears to be writing to ADC channels (from ADC_CHANNEL[3] and up),
even if the channels are being disabled not enabled. Not clear what the
real cause is yet.

Lots of workarounds in code now, none of which appear to have a large impact.